requirements.txt
Scrapy>=2.13,<3
fake-useragent>=1.5
python-dotenv>=1.0
